Atualização do tipo da soma

+++ Exercícios/Ze do Lanche/src/App.java
@@ -12,7 +12,7 @@ public class App {
 
         scanner.close();
 
-        int sumOfBurguersConsumed = 0;
+        float sumOfBurguersConsumed = 0f;
 
         for (int i : consumedBurgers) {
             sumOfBurguersConsumed += i;
